+++ src/frontend/org/voltdb/fault/NodeFailureFault.java
@@ -16,7 +16,6 @@
  */
 package org.voltdb.fault;
 
-import java.util.HashSet;
 import java.util.Set;
 
 public class NodeFailureFault extends VoltFault
@@ -26,7 +25,7 @@ public class NodeFailureFault extends VoltFault
     public static int NODE_FAILURE_INITIATOR = 2;
     public static int NODE_FAILURE_EXECUTION_SITE = 3;
 
-    public NodeFailureFault(int hostId, Set<Integer> failedNonExecSites, String hostname)
+    public NodeFailureFault(int hostId, Set<Long> failedNonExecSites, String hostname)
     {
         super(FaultType.NODE_FAILURE);
         assert(hostname != null);
@@ -40,7 +39,7 @@ public class NodeFailureFault extends VoltFault
         return m_hostId;
     }
 
-    public Set<Integer> getFailedNonExecSites() {
+    public Set<Long> getFailedNonExecSites() {
         return m_failedNonExecSites;
     }
 
@@ -78,6 +77,6 @@ public class NodeFailureFault extends VoltFault
     }
 
     private int m_hostId;
-    private Set<Integer> m_failedNonExecSites;
+    private Set<Long> m_failedNonExecSites;
     private String m_hostname;
 }
